Restore lenses, and copy/paste FakeArtifact into lens_test.go
